// src/AdvancedTechniques/Monopoly/game/monopoly/tiles/tiles.types.ts
export enum TileType {
  Basic = 'basic',
  Property = 'property',
  Action = 'action',
}

export enum ActionType {
  Go = 'go',
  Community = 'community',
  Chance = 'chance',
  Tax = 'tax',
  Arrest = 'arrest',
}

export enum PropertyType {
  Building = 'building',
  Station = 'station',
  Utility = 'utility',
}

export enum BasicType {
  Jail = 'jail',
  Parking = 'parking',
}